Beginning Object code

+++ compiler/src/polyllvm/ast/PseudoLLVM/Expressions/LLVMIntLiteral_c.java
@@ -10,23 +10,24 @@ import polyllvm.ast.PseudoLLVM.LLVMTypes.LLVMTypeNode;
 public class LLVMIntLiteral_c extends LLVMExpr_c implements LLVMIntLiteral {
     private static final long serialVersionUID = SerialVersionUID.generate();
 
-    protected int value;
+    protected long value;
     protected LLVMTypeNode typeNode;
 
-    public LLVMIntLiteral_c(Position pos, int value, LLVMTypeNode tn, Ext ext) {
+    public LLVMIntLiteral_c(Position pos, long value, LLVMTypeNode tn,
+            Ext ext) {
         super(pos, ext);
         this.value = value;
         typeNode = tn;
     }
 
-    protected <N extends LLVMIntLiteral_c> N value(N n, int value) {
+    protected <N extends LLVMIntLiteral_c> N value(N n, long value) {
         if (n.value == value) return n;
         n = copyIfNeeded(n);
         n.value = value;
         return n;
     }
 
-    protected <N extends LLVMIntLiteral_c> N reconstruct(N n, int value) {
+    protected <N extends LLVMIntLiteral_c> N reconstruct(N n, long value) {
         return value(n, value);
     }
 
@@ -37,7 +38,7 @@ public class LLVMIntLiteral_c extends LLVMExpr_c implements LLVMIntLiteral {
 
     @Override
     public String toString() {
-        return Integer.toString(value);
+        return Long.toString(value);
     }
 
     @Override
